Item 7096: Sample Timeline - Server Dev 2
